// my-app/src/components/LibraryProfileEdit/types.ts
export interface LibraryRegistrationFormValues {
  name: string
  country: string
  city: string
  street: string
  number: string
  zip: string
  phone: string
}

export enum LIB_REGISTR_FORM_NAMES {
  NAME = "name",
  COUNTRY = "country",
  CITY = "city",
  STREET = "street",
  NUMBER = "number",
  ZIP = "zip",
  PHONE = "phone",
}